update maxContacts logic in physics collision callbacks. global "maxContacts" is presented as the default upper limit of pairwise contacts, defaults to 100. Geom::maxContacts overrides global setting. Max contacts between pair of geoms is the max of each geom's limit.

+++ server/physics/PhysicsEngine.hh
@@ -183,7 +183,7 @@ namespace gazebo
     /// \brief access functions to set ODE parameters
     public: virtual void SetContactSurfaceLayer(double layer_depth) {}
     /// \brief access functions to set ODE parameters
-    public: virtual void SetMaxContacts(double max_contacts) {}
+    public: virtual void SetMaxContacts(int max_contacts) {}
 
     /// \brief access functions to set ODE parameters
     public: virtual double GetWorldCFM() {return 0;}
@@ -200,7 +200,7 @@ namespace gazebo
     /// \brief access functions to set ODE parameters
     public: virtual double GetContactSurfaceLayer() {return 0;}
     /// \brief access functions to set ODE parameters
-    public: virtual double GetMaxContacts() {return 0;}
+    public: virtual int GetMaxContacts() {return 0;}
 
     /// \brief Setup contact visuals
     public: void SetContactVisual(int num_contacts);
